Add 16 Color support
Based on the version from
<http://saanlima.com/forum/viewtopic.php?p=1793#p1793> which uses a 16
color palette at address -128.

To use, you will have to compile Display.Color.Mod, and then restart the
emulator with the `--color` switch. If the switch is not set, the
emulator behaves as before (except it will use a larger display memory).

Display.Color.Mod contains a hack so that if white=1, white gets copied
from color 15 to 1, and black from 0 to 14. That way, you can avoid
recompiling the whole system. If you recompile the whole system with
white=15, the hack disables itself automatically and all 16 colors can
be used.