Add File Drop support
When a file is dropped on the emulator, it gets copied into the
emulation environment via PCLink. To make this possible when dragged
files are not in the emulator's current directory, the parsing of
PCLink.SND and PCLink.REC is enhanced - when there are two words in the
file, the first word is the file name (as previously) and the second one
is the full path to the file.