Improve FLOOR floating point implementation
The original FLOOR implementation returns incorrect values if the
floating point number is larger than 2^24. While one can argue that
these values might not be exact (as the mantissa is not long enough to
represent every possible value), the default implementation of
Texts.WriteRealFix (which is used for debugging) will often be outside
this range when called with maximum precision of 7, and these incorrect
values can be very confusing.

Note that this workaround will *not* work that easy in Verilog, but
probably it makes the environment more pleasant to use.